refactor: move user deletion hook to hooks directory
